Domset: move nano flann into libs/Math
